requirements.txt
# Core Dependencies
streamlit==1.31.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
pydantic>=2.0.0

# UI/Visualization
altair>=5.0.0
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dateutil>=2.8.2
pillow>=10.0.0
